refactor: extract to directory
