+++ auth/auth_contract.ts
